+++ mrlib/round_float.go
@@ -18,15 +18,21 @@ func RoundFloat(x float64, precision int) float64 {
 
 // RoundFloat2 - возвращает RoundFloat с точностью 2.
 func RoundFloat2(x float64) float64 {
-	return RoundFloat(x, 2)
+	const precision = 2
+
+	return RoundFloat(x, precision)
 }
 
 // RoundFloat4 - возвращает RoundFloat с точностью 4.
 func RoundFloat4(x float64) float64 {
-	return RoundFloat(x, 4)
+	const precision = 4
+
+	return RoundFloat(x, precision)
 }
 
 // RoundFloat8 - возвращает RoundFloat с точностью 8.
 func RoundFloat8(x float64) float64 {
-	return RoundFloat(x, 8)
+	const precision = 8
+
+	return RoundFloat(x, precision)
 }