+++ mrtype/types.go
@@ -12,10 +12,16 @@ type (
 	// KeyString - строковый ID записи.
 	KeyString string
 
-	// RangeInt64 - comment struct.
+	// RangeInt64 - целочисленный интервал [Min, Max].
 	RangeInt64 struct {
 		Min int64
 		Max int64
+	}
+
+	// RangeFloat64 - вещественный интервал [Min, Max].
+	RangeFloat64 struct {
+		Min float64
+		Max float64
 	}
 
 	// PageParams - параметры для выборки части списка элементов.
@@ -30,3 +36,12 @@ type (
 		Direction mrenum.SortDirection // sortDirection
 	}
 )
+
+// Transform - преобразовывает в RangeFloat64 с умножением полей на указанный коэффициент
+// (для приведения к необходимой ед. измерения).
+func (r RangeInt64) Transform(coefficient float64) RangeFloat64 {
+	return RangeFloat64{
+		Min: float64(r.Min) * coefficient,
+		Max: float64(r.Max) * coefficient,
+	}
+}